whitelist-api/requirements.txt
werkzeug==2.1.2
flask
jsonify
flake8
mongoengine
flask_cors
pymongo
flask_pymongo
python-dotenv
flask-restx